This time I happened to be rushing.
But I still have some unique stuff upon here! The appearance of Void Bundle is inspired to that similar thing in the movie Inside Out,with an old and worn-up look. I made some really-bad-but-a-bit-better pigment designs,a more scientific Shooting Star design,a better sign of laziness and some invaluable stuff.
Also,I tried to make Bedrock Sword to be rough,to show the toughness and craft difficulty of the bedrock. I also planned to make those colorful  spore blossoms unique to each other,just like those colorful trees!