Make Modonomicon build in Yarn
Requires only using the fabric jar, as the common jar requires Mojmap in order to be used. The exclude prevents it from being included.
Also the guidebook calls can only be made on the clients render thread. This is different from Patchouli, which seemed to send a S2C packet to open the correct page. A small refactor was made in order to test this.
Guidebook is still not openable, due to the main entry not existing.

Signed-off-by: Noaaan <noaaan@hotmail.com>