Additional Network error state fix
Fix for creating networks with nodes A, B, C, with network inheriting node A's UUID, disconnecting node A (retaining network UUID), and connecting node A to node D, leading to node A and D being connected to network with A's UUID despite no connections, leading to an incorrect network state. This means recreating a network with the same nodes may have a different colour.